refractor

+++ conftest.py
@@ -10,7 +10,7 @@ def pytest_cmdline_preparse(args):
     if "xdist" in sys.modules:  # pytest-xdist plugin
         import multiprocessing
 
-        num = max(multiprocessing.cpu_count() / 2, 1)
+        num = int(max(multiprocessing.cpu_count() / 2, 1))
         args[:] = ["-n", str(num)] + args
 
     # add mypy option if not pypy
